Msgs Seen Imple

+++ app/src/main/java/com/mardomsara/social/models/tables/Message.java
@@ -81,7 +81,7 @@ public class Message {
     public int ToPush = 0;
 
 	@Column(defaultExpr = "0", helpers = Column.Helpers.CONDITION_EQ)
-	public int ISeenTime = 0;
+	public long ISeenTime = 0;
 
     @Column(defaultExpr = "0" ,helpers = Column.Helpers.CONDITION_EQ)
     public int MediaStatus=0;
